[CHANGE] Moved all classes to a package and renamed files
