--- src/main/scala/error.scala
@@ -1,3 +1,0 @@
-package recast.error {
-  case class RecastError(message: String) extends Exception(message)
-}
