+++ src/main/scala/recast/RecastError.scala
@@ -1,0 +1,3 @@
+package recast
+
+case class RecastError(message: String) extends Exception(message)
